add: 0912 bucked ckpts
