+++ mujoco_playground/mujoco_playground/_src/locomotion/__init__.py
@@ -45,6 +45,8 @@ from hector_pg.wbc import wbc as hector_wbc
 from hector_pg.wbc import randomize as hector_wbc_randomize
 from hector_pg.joystick import joystick as hector_joystick
 from hector_pg.joystick import randomize as hector_joystick_randomize
+from hector_pg.joystick_ilab import joystick as hector_joystickilab
+from hector_pg.joystick_ilab import randomize as hector_joystickilab_randomize
 
 
 _envs = {
@@ -96,6 +98,9 @@ _envs = {
     "HectorJoystickFlatTerrain": functools.partial(
         hector_joystick.Joystick, task="flat_terrain"
     ),
+    "HectorJoystickFlatTerrainIlab": functools.partial(
+        hector_joystickilab.Joystick, task="flat_terrain"
+    ),
 }
 
 _cfgs = {
@@ -125,6 +130,7 @@ _cfgs = {
     
     "HectorWBCFlatTerrain":hector_wbc.default_config,
     "HectorJoystickFlatTerrain":hector_joystick.default_config,
+    "HectorJoystickFlatTerrainIlab":hector_joystickilab.default_config,
 }
 
 _randomizer = {
@@ -146,6 +152,7 @@ _randomizer = {
     
     "HectorWBCFlatTerrain":hector_wbc_randomize.domain_randomize,
     "HectorJoystickFlatTerrain":hector_joystick_randomize.domain_randomize,
+    "HectorJoystickFlatTerrainIlab":hector_joystickilab_randomize.domain_randomize,
 
 }
