+++ mujoco_playground/mujoco_playground/config/locomotion_params.py
@@ -149,7 +149,8 @@ def brax_ppo_config(env_name: str) -> config_dict.ConfigDict:
 
   elif env_name in (
       "HectorWBCFlatTerrain",
-      "HectorJoystickFlatTerrain"
+      "HectorJoystickFlatTerrain",
+      "HectorJoystickFlatTerrainIlab",
   ):
     rl_config.num_timesteps = 150_000_000
     rl_config.num_evals = 20
@@ -157,7 +158,7 @@ def brax_ppo_config(env_name: str) -> config_dict.ConfigDict:
     rl_config.clipping_epsilon = 0.2
     rl_config.num_resets_per_eval = 1
     rl_config.entropy_cost = 0.005
-    rl_config.learning_rate=1.2e-4
+    rl_config.learning_rate=3e-4
     rl_config.network_factory = config_dict.create(
         policy_hidden_layer_sizes=(1024, 512, 256),
         value_hidden_layer_sizes=(1024, 512, 256),
