+++ src/hector_pg/joystick_ilab/joystick.py
@@ -21,13 +21,15 @@ from hector_pg.utils import ankle_decouple
 from dataclasses import dataclass
 from typing import Callable
 
+from . import rewards
+
 def default_config() -> config_dict.ConfigDict:
   return config_dict.create(
       ctrl_dt=0.02,
       sim_dt=0.002,
       episode_length=1000,
       action_repeat=1,
-      action_scale=0.75,
+      action_scale=0.50,
       history_len=1,
       soft_joint_pos_limit_factor=0.95,
       # OBS size
@@ -53,18 +55,14 @@ def default_config() -> config_dict.ConfigDict:
               # --- Tracking related rewards ---
               tracking_lin_vel=2.0, #2.0.
               tracking_ang_vel=1.5, #1.5
-              #tracking_vel_hard=0.0,
-              #tracking_body_height=0.0,
-              #tracking_body_euler=0.0,
-              #tracking_arm=0.0,
               # --- Base related rewards ---
-              lin_vel_z=-0.15,
-              ang_vel_xy=-0.25,#-0.25,
-              orientation=1.0,
+              lin_vel_z=-0.75,
+              ang_vel_xy=-0.75,
+              #orientation=0.0,
               # --- Energy related rewards ---
-              energy=-0.0,
+              #energy=-0.0,
               smoothness=-0.005,
-              contact_force=-0.0,
+              #contact_force=-0.0,
               #dof_acc = -0.0, #-1e-7,
               #dof_vel = -0.0, #-1e-4,
               # --- Feet related rewards ---
@@ -72,24 +70,24 @@ def default_config() -> config_dict.ConfigDict:
               feet_height=2.0,
               feet_slip=-0.5,
               undesired_contact=-3.0,
-              feet_upright=-0.25,
-              feet_dist=-0.0,#-5e4,
+              #feet_upright=-0.25,
+              feet_dist=-4e3,
               # --- Other rewards ---
               alive=0.5,
               termination=-1.0,
               #stand_still=-0.0, # -1.0
               # --- Pose related rewards ---
               #joint_deviation_knee=-0.0,
-              joint_deviation_hip=-0.25,
-              dof_pos_limits=-0.25,
-              pose=-0.5,
+              #joint_deviation_hip=-0.0,
+              #dof_pos_limits=-0.0,
+              pose=-1.0,
           ),
           max_foot_height=0.10,
           max_contact_force=250.0,
           # Force threshold that holds as contact
           feet_f_contact = 5.0,
           # Desired airtime within phase (1.0 scale)
-          airtime = 0.45, #0.45
+          airtime = 0.35, #0.45
           # In what precentage control will be ruleout
           default_p = 0.1,
       ),
@@ -107,12 +105,6 @@ def default_config() -> config_dict.ConfigDict:
       f_dist_range=[0.08, 0.4],
       # Default body height
       body_height_default=0.55,
-
-      delay = config_dict.create(
-          min_delay_step=0,
-          max_delay_step=10,
-          hist_len=10,
-      ),
   
       impl="jax", # "jax" or "warp"
       nconmax=8 * 8192,
@@ -127,7 +119,7 @@ class RewardTerm:
   func: Callable[..., jax.Array]
   
 class Joystick(hector_base.HectorEnv):
-  """Track a Whole body control command."""
+  """Track a Joystick command."""
 
   def __init__(
       self,
@@ -157,7 +149,7 @@ class Joystick(hector_base.HectorEnv):
     
     # Print joint limits
     joint_names = [self.mj_model.joint(i).name for i in range(1, self.mj_model.njnt)]
-    for name, low, high in zip(joint_names, self._lowers, self._uppers): print(f"{name}: {low:.2f} to {high:.2f}")
+    #for name, low, high in zip(joint_names, self._lowers, self._uppers): print(f"{name}: {low:.2f} to {high:.2f}")
 
     hip_indices = []
     # To keep hip yaw, roll angle near default angle here
@@ -176,8 +168,8 @@ class Joystick(hector_base.HectorEnv):
 
     # fmt: off
     self._weights = jp.array([
-        0.7, 0.7, 0.01, 0.01, 0.01,  # left leg.
-        0.7, 0.7, 0.01, 0.01, 0.01,  # right leg. # 0.5
+        0.75, 0.2, 0.05, 0.05, 0.05,  # left leg.
+        0.75, 0.2, 0.05, 0.05, 0.05,  # right leg. # 0.5
         0.5, 0.5, 0.5, 0.5,   # left arm
         0.5, 0.5, 0.5, 0.5,   # right arm
     ])
@@ -223,34 +215,34 @@ class Joystick(hector_base.HectorEnv):
     # Init reward terms
     reward_function_mapping = {
         # --- Tracking rewards ---
-        'tracking_lin_vel': self._reward_tracking_lin_vel,
-        'tracking_ang_vel': self._reward_tracking_ang_vel,
+        'tracking_lin_vel': rewards._reward_tracking_lin_vel,
+        'tracking_ang_vel': rewards._reward_tracking_ang_vel,
         #'tracking_vel_hard': self._reward_tracking_vel_hard,
         #'tracking_body_height': self._reward_tracking_body_height,
         # --- Stay balanced ---
-        'lin_vel_z': self._cost_lin_vel_z,
-        'ang_vel_xy': self._cost_ang_vel_xy,
-        'orientation': self._reward_base_orientation,
+        'lin_vel_z': rewards._cost_lin_vel_z,
+        'ang_vel_xy': rewards._cost_ang_vel_xy,
+        #'orientation': rewards._reward_base_orientation,
         # --- Energy terms ---
-        'energy': self._cost_energy,
-        'smoothness': self._cost_smoothness,
+        #'energy': rewards._cost_energy,
+        'smoothness': rewards._cost_smoothness,
         #'dof_acc': self._cost_dof_acc,
         #'dof_vel': self._cost_dof_vel,
-        'contact_force': self._cost_contact_force,
+        'contact_force': rewards._cost_contact_force,
         # --- Gait shaping ---
-        'feet_height': self._reward_feet_height,
+        'feet_height': rewards._reward_feet_height,
         #'feet_air_time': self._reward_feet_air_time,
-        'feet_slip': self._cost_feet_slip,
-        'undesired_contact': self._cost_undesired_contact_phase,
-        'feet_upright': self._cost_feet_upright,
-        'feet_dist': self._cost_feet_dist,
+        'feet_slip': rewards._cost_feet_slip,
+        'undesired_contact': rewards._cost_undesired_contact_phase,
+        #'feet_upright': rewards._cost_feet_upright,
+        'feet_dist': rewards._cost_feet_dist,
         # --- Alive ---
-        'alive': self._reward_alive,
-        'termination': self._cost_termination,
+        'alive': rewards._reward_alive,
+        'termination': rewards._cost_termination,
         #'stand_still': self._cost_stand_still,
         # --- Others ---
-        'dof_pos_limits': self._cost_joint_pos_limits,
-        'pose': self._cost_pose,
+        #'dof_pos_limits': rewards._cost_joint_pos_limits,
+        'pose': rewards._cost_pose,
       # Add other rewards here as we create them
     }
 
@@ -631,148 +623,10 @@ class Joystick(hector_base.HectorEnv):
     }
     rewards = {}
     for term in self._reward_terms:
-        reward_value = term.func(context)
+        reward_value = term.func(self, context)
         rewards[term.name] = reward_value # Store unscaled value for metrics
     return rewards
 
-
-  # Tracking rewards.
-  def _reward_tracking_lin_vel(self, context: Dict[str, Any]) -> jax.Array:
-    commands = context['command']
-    local_vel = context['local_linvel']
-    sigma = 0.25
-    lin_vel_error = jp.sum(jp.square(commands[:2] - local_vel[:2]))
-    return jp.exp(-lin_vel_error / sigma)
-
-  # Yaw tracking reward
-  def _reward_tracking_ang_vel(self, context: Dict[str, Any]) -> jax.Array:
-    commands = context['command']
-    ang_vel = context['gyro']
-    sigma = 0.25
-    ang_vel_error = jp.square(commands[2] - ang_vel[2])
-    return jp.exp(-ang_vel_error / sigma)
-  
-  # Base-related rewards and penalties
-  def _cost_lin_vel_z(self, context: Dict[str, Any]) -> jax.Array:
-    return jp.square(context['global_linvel'][2])
-
-  def _cost_ang_vel_xy(self, context: Dict[str, Any]) -> jax.Array:
-    return jp.sum(jp.square(context['global_angvel'][:2]))
-
-  def _reward_base_orientation(self, context: Dict[str, Any]) -> jax.Array:
-    torso_zaxis = context['torso_zaxis']
-    sigma = 0.25
-    err = jp.sum(jp.square(torso_zaxis[:2]))
-    return jp.exp(-err/sigma)
-
-  # Energy related rewards.
-  def _cost_energy(self, context: Dict[str, Any]) -> jax.Array:
-    qvel = context['qvel']
-    qfrc = context['act_frc']
-    return jp.sum(jp.abs(qvel) * jp.abs(qfrc))
-  
-  def _cost_smoothness(self, context: Dict[str, Any]) -> jax.Array:
-    act = context['action']
-    last_act = context['last_act']
-    last_last_act = context['last_last_act']
-    c1 = jp.sum(jp.square(act - last_act))
-    c2 = jp.sum(jp.square(act - 2*last_act + last_last_act))
-    return (c1+c2)
-  
-  def _cost_dof_acc(self, context: Dict[str, Any]) -> jax.Array:
-    return jp.sum(jp.square(context['qacc']))
-
-  def _cost_dof_vel(self, context: Dict[str, Any]) -> jax.Array:
-    return jp.sum(jp.square(context['qvel']))
-
-  # Others
-  def _cost_joint_pos_limits(self, context: Dict[str, Any]) -> jax.Array:
-    qpos = context['q']
-    out_of_limits = -jp.clip(qpos - self._soft_lowers, None, 0.0)
-    out_of_limits += jp.clip(qpos - self._soft_uppers, 0.0, None)
-    return jp.sum(out_of_limits)
-
-  def _cost_termination(self, context: Dict[str, Any]) -> jax.Array:
-    return context['done']
-
-  def _reward_alive(self, context: Dict[str, Any]) -> jax.Array:
-    return jp.array(1.0)
-
-  # Pose-related rewards.
-  def _cost_joint_deviation_hip(self, context: Dict[str, Any]) -> jax.Array:
-    qpos = context['q']
-    cost = jp.sum(jp.abs(qpos[self._hip_indices] - self._default_pose[self._hip_indices]))
-    cost *= jp.abs(context['command'][1]) > 0.1 # Maskout if there is large vy command
-    return cost
-
-  def _cost_joint_deviation_knee(self, context: Dict[str, Any]) -> jax.Array:
-    qpos = context['q']
-    err = qpos[self._knee_indices] - self._default_pose[self._knee_indices]
-    return jp.sum(jp.abs(err))
-    
-  def _cost_contact_force(self, context: Dict[str, Any]) -> jax.Array:
-    data = context['data']
-    max_fz = context['max_fz']
-    l_f = mjx_env.get_sensor_data(self.mj_model, data, "left_foot_force")
-    r_f = mjx_env.get_sensor_data(self.mj_model, data, "right_foot_force")
-    l_fz = l_f[2]
-    r_fz = r_f[2]
-    return jp.clip(jp.abs(l_fz)+jp.abs(r_fz), 0.0, 200.0)
-
-  def _cost_pose(self, context: Dict[str, Any]) -> jax.Array:
-    qpos = context['q']
-    return jp.sum(jp.square(qpos - self._default_pose) * self._weights)
-
-  # Feet related rewards.
-  def _cost_feet_slip(self, context: Dict[str, Any]) -> jax.Array:
-    data = context['data']
-    contact = context['contact']
-    feet_vel = data.sensordata[self._foot_linvel_sensor_adr]  # (2, 3)
-    v_tan = jp.linalg.norm(feet_vel[..., :2], axis=-1)        # (2,)
-    # Penalize slip only when that foot is in contact
-    return jp.sum(jp.where(contact, v_tan, 0.0))
-
-  def _reward_feet_height(self, context: Dict[str, Any]) -> jax.Array:
-    sigma = 0.0004
-    rz = gait.get_rz_phase(context['phase'],
-                           swing_height=context['max_foot_height'],
-                           airtime=context['airtime'])    
-    err = jp.clip(rz - context['p_fz'], min=0.0)
-    squared_err = jp.square(err)
-    rew_per_foot = jp.exp(-squared_err/sigma)
-    # Rule out zero commands, stand still
-    cmd_norm = jp.linalg.norm(context['command'][0:3])
-    rew_swing = rew_per_foot * jp.logical_not(context['contact']) * (cmd_norm>0.1)
-    return jp.sum(rew_swing)
-  
-  def _cost_feet_upright(self, context: Dict[str, Any]) -> jax.Array:
-    z_fz = context['zaxis_fz']
-    c_l = jp.sum(jp.square(z_fz[0])) # Only care about x axis projection
-    c_r = jp.sum(jp.square(z_fz[3]))
-    return c_l+c_r
-  
-  def _cost_feet_dist(self, context: Dict[str, Any]) -> jax.Array:
-    p_f = context['data'].site_xpos[self._feet_site_id]
-    dmin, dmax = context['f_dist_range'] # min max of target feet distance
-    dist = jp.linalg.norm(p_f[0,:2]-p_f[1,:2])
-    under = jp.maximum(dmin - dist, 0.0)
-    over  = jp.maximum(dist - dmax, 0.0)
-    violation = under + over
-    return violation*violation
-  
-  def _cost_stand_still(self, context: Dict[str, Any]) -> jax.Array:
-    commands = context['command']
-    qpos = context['q']
-    cmd_norm_twist = jp.linalg.norm(commands[0:3])
-    cmd_norm_track = jp.linalg.norm(commands[3:7]-jp.array([context['tar_body_height'], 0.0, 0.0, 0.0]))
-    enable = (cmd_norm_twist<0.1) & (cmd_norm_track<0.25)
-    return jp.sum(jp.abs(qpos[0:10] - self._default_pose[0:10])) * enable
-
-  def _cost_undesired_contact_phase(self, context: Dict[str, Any]) -> jax.Array:
-    contact = context['contact']
-    desired_contact = context['desired_contact']
-    mismatch = jp.not_equal(contact, desired_contact) 
-    return jp.sum(mismatch.astype(jp.float32))
   
   # Sample in command space, command dim defined here
   def sample_command(self, rng: jax.Array) -> jax.Array:
