+++ src/train/train_jax_ppo.py
@@ -64,7 +64,7 @@ warnings.filterwarnings("ignore", category=UserWarning, module="absl")
 
 _ENV_NAME = flags.DEFINE_string(
     "env_name",
-    "HectorJoystickFlatTerrain",
+    "HectorJoystickFlatTerrainIlab",
     f"Name of the environment. One of {', '.join(registry.ALL_ENVS)}",
 )
 _IMPL = flags.DEFINE_enum("impl", "jax", ["jax", "warp"], "MJX implementation")
